+++ docs/conf.py
@@ -1,3 +1,0 @@
-from fairdm_docs.conf import *  # noqa: F401, F403
-
-# -- Overrides -----------------------------------------------------